requirements.txt
streamlit==1.28.1
oracledb==1.4.2
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3